requirements.txt
pandas
numpy
scikit-learn
nltk
matplotlib
seaborn
jupyter
PyJWT
